new dialogs take you from sign up to login
